Full matching (not just embedded) now works in central Hijing events.
g4hit + Cluster + track matching executes in 1200 seconds
including the hit matching (not optimized yet) increases the budget to 3800seconds.

In semi central events the CPU usage decreased from 2700sec to 465sec.

Most performance inceases were obtained by presorting the clustermap
in layer and z to reduce the search range and by caching matching results.
A lot of time is spent accessing the clustermap. Access to clusters in the large 
clustermap is apparently quite inefficient.